azure-functions/requirements.txt
# Azure Functions MCP Server dependencies
azure-functions>=1.18.0
azure-identity>=1.15.0
httpx>=0.27.0
pydantic>=2.0
